requirements.txt
# 基础依赖
selenium==4.10.0
selenium-wire==5.1.0
requests==2.28.2
Pillow==9.5.0
SQLAlchemy==2.0.16

# Celery相关
celery==5.3.1
redis==4.5.5

# Web服务
Flask==2.3.2
gunicorn==20.1.0

# 可能需要的其他包
cryptography==41.0.1
python-dotenv==1.0.0 